Add a build to check code formatting. (#51)
Configure one of the builds to run clang-format on all the code
and break if the formatting is wrong.  That will force us to
follow the format in all our code.

The implementation is a bit wonky because of #49, but works.
Also fixed a few formatting errors so the build would pass.

This fixes #17.