Fix sanitizer builds and test them.
In this commit I intentionally enable tests that force every type
of sanitizer error.  A future commit will disable those tests, but
running the builds now will validate the script and show that it
detects the right things.

We disable the MemorySanitizer tests because they require all
libraries (including libstdc++ or libc++) to be compiled with
the sanitizer turned on.  See #62 for details.

Also fixed a number of bugs in the scripts that I somehow missed
in earlier.