Create scripts to generate and upload documentation.
This is part of the fixes for #18.  It is incomplete because it
does not generate the documentation in the format used by other
google-cloud-* repositories.  We will need to convert the doxygen
output into the format we use in those other repositories (gulp
I believe).

The upload to github.io is only enabled if GH_TOKEN is set, we
should consult with the owners of GoogleCloudPlatform.github.io
before posting the docs there.

I think this partial implementation is still useful because (a) it
will detect documentation errors, and (b) it will post to my fork,
where at least we can see it it looks right.