Fix sanitizer builds and test them. (#65)
I first intentionally enabled tests that force every type of sanitizer error; [this build](https://travis-ci.org/coryan/google-cloud-cpp/builds/310318481) shows the results of running that version. I then disabled those tests prior to committing this change.

We can enable them in the future when we change the scripts; to enable them when running locally, use:

```console
$ cmake -DFORCE_SANITIZER_ERRORS=yes
```

We disable the MemorySanitizer tests because they require all
libraries (including libstdc++ or libc++) to be compiled with
the sanitizer turned on.  See #62 for details.

Also fixed a number of bugs in the scripts that I somehow missed earlier.

This fixes #9 (again), because the first PR was broken.