Introduce policy to decide if mutations are idempotent. (#64)
Introduce policy to decide if mutations are idempotent.

Mutations that are not idempotent are not retried, regardless of whether we have time to do so.  Introduced both a safe policy for the common case, and a policy to retry all mutations for the less common, but useful case.

Many thanks to the reviewers, @mbrukman and @garye.

This fixes #13.