Enable scan build (#157)
Run Clang static analyzer against installed gRPC.

Compiling with the Clang static analyzer is fairly slow, and
can take longer than the maximum time allowed by Travis if gRPC
is part of the build.  In this change we compile and install
gRPC first, and then compile the code against it.

That seems to be the only way to stop cmake from using the static
analyzer everywhere.  And frankly it is a good thing that we test
the "package" version of gRPC too.

I think this could be a fix for #10, but I will let @mbrukman decide.