Fix document upload after #138. (#154)
The documents are generated inside a docker image, in a previous
change (#138), the path for the generated docs was changed. This
script should have been changed at the same time.

While testing I found out that there was another tiny problem, the
documents were never uploaded because the change detection was
broken.  Fix that too <blush>.