Provide custom CMakeLists.txt for Abseil.
Abseil does not install its libraries or headers, and forces the
libraries to be static.  With this change we install the libraries
ourselves, and allow the user to decide if static or shared
libraries work better for them.  We also fix some problems
with Abseil requiring gtest and gmock targets even when testing
is disabled.