Fix broken build with installed gRPC. (#180)
Mainly we cannot use headers that are not installed by gRPC, so I removed their use in a couple of tests.  I also discovered that Fedora:27 is ongoing changes to the clang-format version they have during this weekend, so half the mirrors get clang-format-4.0 and the other half get clang-format-5.0 which results in broken builds because the formatting is different in a few files.

And finally I changed the build matrix to prevent these problems in the future.