Refactor bigtable::AdminClient and bigtable::DataClient. (#170)
* Refactor bigtable::AdminClient and bigtable::DataClient.

There was a certain amount of duplication in these classes. This
change fixes that.  As a side-effect, it fixes #107 too.

This change does not attempt to *use* both classes the same way,
though we should, I do not want to introduce more changes in a
PR that basically refactors things.

Use CreateDefault*Client instead of some with Default and some
without.  Also use Default*Client instead of Simple*Client for
one and Default*Client for the other.  Thanks to @garye for
noticing.