Benchmark ReadRow() and Apply() latency.
This fixes #189.  The benchmark can run against production
instances (the default), against the Cloud Bigtable Emulator (if
BIGTABLE_EMULATOR_HOST is set), or against an embedded gRPC
server.

The benchmark outputs the results in human readable form (as
comments starting with '#'), as well as CSV lines.  This makes it
possible to easily load the data into analysis tools, such as R or
SciPy, or Cloud DataLab).