Use target_include_directories() for the libraries. (#201)
Modern CMake prefers target_include_directories() over
include_directories().  The latter changes the include path
globally, the former changes the include path for targets that
depend on a given library.
Also create bigtable::* aliases for the libraries as this is the
preferred style (it seems) for exported CMake targets.

Thanks to @DanielMahu for spotting the typo in one of the target names.