Helper classes to initialize PRNGs. (#200)
The benchmarks in #199, #196, and #189 will require good
pseudo-random number generators (PRNGs).  Unfortunately C++11
PRNGs are notoriously hard to initialize, and Abseil does
not offer a solution either.  We also introduce some
helper functions to create random strings in the benchmarks.

Introduce dependency on absl::time, not strictly needed in this PR,
but will soon, and that will need Win64 builds anyway.