Helper class to create an embedded Bigtable server. (#212)
The benchmarks in #199, #196, and #189 will need to run against
the production Bigtable server, against the Cloud Bigtable
Emulator, or against a server running in the same process as the
benchmark.  This class make it easy to create such an embedded
server.

Implement unit test for embedded server.  I realized late that I
needed one, and found many bugs thanks to it.

Use IPv4 in embedded server benchmarks. This is because
Travis has a weird configuration, the default
address for localhost is IPv6 (::1), but Travis has disabled
local IPv6:

https://github.com/travis-ci/travis-ci/issues/5200
https://github.com/travis-ci/travis-ci/issues/4964