Update to current gRPC branch (v1.9.x) (#213)
Fix #211.  It was a bug in gRPC so we move to a more recent
version.  I tried `master` but it does not compile on Ubuntu 17.10.
Luckily, v1.9.0 has just been released so we can use that.  Add
a small test to reproduce the problem in the future, it takes hours
to run, so it is not part of the CI tests.  The test might be useful
next time we find a problem in gRPC.