Support Abseil as an installed library.
This commit implements the design described in
doc/managing-dependencies-in-cmake.md.  All libraries are
supported as either submodules, packages installed with CMake
support (i.e., discoverable using find_package()), or using
pkg-config.

This change also simplifies the code because now we can use the
same name (e.g. gRPC::grpc++) for a library regardless of how it
is discovered.  When the library is discovered we create the
correct name for it if necessary.