Do not return absl::string_view objects. (#216)
They are not that useful because one has to explicitly convert
them to std::string to use them with most C++ classes, for
example, the std::string constructor does not accept
absl::string_view (in C++17 it would accept std::string_view, but
that is not the version we use).  Same problems with
std::istringstream, or with protos.

This is part of the fixes for #32.