Do not use CMake-3.9 features.
I was accidentally using CMake-3.9 features (IMPORTED_TARGET in
pkg_check_modules()).  We support 3.5 and higher, so change the
code to not depend on this feature.  Specifically, I had to
manually convert the command-line flags found by
pkg_check_modules() into properties that can be used in an IMPORTED
library.