Upgrade Google Cloud SDK after installing.
Upgrading after installing seems to work, apparently the bug
in v188 is not triggered because the install of v187 "remembers"
some things.

This could work as a fix for #224, always install a known-good
version and upgrade from that version.  Every so often we would
need to update what the known-good version is, but that can be
done without disrupting the current builds.