Got the tests to pass without exceptions.
Generally, we cannot use EXPECT_NO_THROW() in tests.  In some
cases we can use EXPECT_DEATH_*() to verify the library aborts,
but that does not interact well with mocks: the expectations fail
when they run on a death test because those execute in a child
process.
Finally avoid try/catch blocks in main when possible.