Enable a build with exceptions disabled. (#237)
This is part of the fixes for #35.  The changes are (I think) the minimal set of changes to get a build to compile and pass the tests without exception support, but even so, it is a large PR.  There are several changes:

- Introduces a configuration flag in CMake to disable exceptions.
- It creates an automated build using that flag.
- It changes the code to raise exceptions through a delegate function, an idea stolen from Abseil, which aborts on exceptions.  A future PR we will add APIs that return errors instead of crashing.
- It changes the tests to use `EXPECT_DEATH()` when exceptions are disabled instead of EXPECT_THROW().

Some tests are disabled when exceptions are unavailable, and mocking is complicated with `EXPECT_DEATH()` because the side-effects in the mocks are not observable (they happen in a separate process).