Do not use absl::optional<> (#264)
This fixes #252.  Use a small class to represent optional rows. 
Thanks to @richdutton for the suggestion.  The previous code left
"row" and "has_row" as two disconnected things.  On the other
hand, implementing a full std::optional<> replacement is overkill.
So we implement a small OptionalRow class to do the work.