Implement connection pooling. (#272)
This fixes #204.  With this change the application can control
the size of the pool associated with each client using
`ClientOptions::set_connection_pool_size()`.  The default size is 4,
which happens to match the value in Go. Note that by default the
channels created by different clients are shared, but the application
can override that using `ClientOptions::set_connection_pool_name()`.
We expect that to be rarely used.