Complete implementation of install target.
This fixes #24, it contains an example to test 'make install' as
part of the continuous integration build.  It also fixes #44, it
creates and installs the pkg-config support files.  Finally it
fixes #45, it creates and installs the CMake configuration files
to use find_package() when the libraries are installed.

Apologies for making so many fixes in one commit, but otherwise
the commits need to create pkg-config files (or CMake config
files) that are not tested, or I need to introduce tests that
have everything hardcoded without the config files.