Do not force -Wall -Werror on the user.
We want to enable -Wall -Werror on our code, but we do not want
to force it on the user, nor do we want to force it on our
dependencies.  I needed to make the change now to workaround the
build problems referenced in:

https://github.com/grpc/grpc/issues/14511

But they are good changes in general too.