Fix bug in scan throughput benchmark. (#299)
I made a mistake in the scan throughput benchmark, swapping the
scan_size and table_size parameters. Effectively this meant all
the tests were using the same scan size, and it was absurdly
large (the full table!).  Of course that resulted in fantastic
performance results, but for a different thing than we wanted to
mention.

I changed the function signatures to make it less likely that one
can swap the parameters.