Refactor AdminClient to support multiple services. (#444)
This is a large change. It is part of the fixes for #433 and
for #418. The current implementation of bigtable::*Client makes
it really hard to write tests that use multiple services, and
when using long running operations you need to access at least
two services.

The good news is that the tests and some of the helpers get
easier to write, and this does not affect any of the APIs exposed
to the user.