Support Bazel-based builds.
This fixes #334. It creates Bazel BUILD and WORKSPACE files to
compile the code.  The files are automatically kept in sync
with the CMake files: our CMake files generate (most of) the
contents for the Bazel files.  This also introduces a Bazel CI
build, and tests to verify the Bazel files are usable when
embedded from another project.