Run docker builds with the user id.
By default Docker containers run as root.  Our build scripts
execute inside the container, but mount the user's workspace to
read the code and to output build artifacts (both `.o` files and
final outputs). That is annoying because any files need to be
removed with `sudo`.  With this change the docker script uses the
user's uid for most builds.

The only exception are the TEST_INSTALL=yes builds, where it is
useful to run as a user with permissions to write in `/usr/local/`.
Note that this is the `/usr/local` directories inside the Docker
image, so none of these changes leak to the user's workstation.