Format with clang-format-6.0 on ubuntu:18.04. (#597)
Moving the clang-format build to ubuntu:18.04 allows us to merge
he clang-tidy and clang-format builds, saving a build for more
useful stuff.  Part of the fixes for #19.

If you are curious, the newer clang-format enforces the comments
when closing a namespace (`}  // namespace foo`), highlighted
that we were using `[[noreturn]]` attributes on a function
definition (where they have no effect), and has very different
opinions about formatting complex lambdas.

As usual I do not care about what clang-format produces, just
happy it is close enough to what a human would type, and we can
reproducibly run it.