Preserve error messages in grpc::Status. (#604)
On idempotent operations we were not preserving the error message
returned by the server, that left the users with poor errors.

On non-idempotent operations we were not inserting the library
call that produce the error, which is also not great.

Use testing::HasSubstr(), far more readable, to check that strings have
a given substring. Thanks to @mbrukman for the suggestion.