Speed up build+install of dependencies with ccache.
Some builds need the dependencies (gRPC, protobuf, c-ares)
pre-installed.  We used to install the dependencies as part of the
Docker image creation, with this change the dependencies are
installed by the Travis script.  That allows us to use ccache
to speed up the build+install of these dependencies, just like
we do for the main code.