Use ccache for SCAN_BUILD and TEST_INSTALL builds.
Enable ccache in these builds, they are slightly different from
the rest of the builds in that they run as root within the
container, so the cache directories are different.

For the SCAN_BUILD=yes builds I could not figure out how to use
both scan-build and ccache at the same time.  Both replace CXX
and CC, so configuring that is tricky.  I was able to use
ccache for installing the dependencies (which do not require
static analysis), and then disabled ccache for running the static
analysis on our code. We can improve on this later.