Take advantage of ccache on builds that install the dependencies (#614)
Some builds need the dependencies (gRPC, protobuf, c-ares)
pre-installed.  We used to install the dependencies as part of the
Docker image creation, with this change the dependencies are
installed by the Travis script.  That allows us to use ccache
to speed up the build+install of these dependencies, just like
we do for the main code.

Enable ccache in these builds, they are slightly different from
the rest of the builds in that they run as root within the
container, so the cache directories are different.

For the SCAN_BUILD=yes builds I could not figure out how to use
both scan-build and ccache at the same time.  Both replace CXX
and CC, so configuring that is tricky.  I was able to use
ccache for installing the dependencies (which do not require
static analysis), and then disabled ccache for running the static
analysis on our code. We can improve on this later.

Make the logs a bit more readable.

The builds that install their dependencies (TEST_INSTALL,
SCAN_BUILD), do not need to download the gRPC submodule,
which saves time.

Thanks to @mbrukman and @paprat for their comments.