Run integration tests against a fake.
Our code coverage results are poor because we cannot run most of
the code in the coverage build. This change (more of a proposal
really), runs the code against a super simplistic fake of GCS.
While I expect that fake to grow in complexity, and to need
refactoring over time, I do not expect it to be much of a burden.

To make it work I also introduce insecure credentials, and an
environment variable to configure which endpoint (production vs.
the fake) is used by the library.