Include analysis reports in the Travis logs.
Some of our analysis tools (the ABI compliance checker and the
Clang static analyzer) present their results as HTML reports.
Include the output of these reports as text when they are present.

The dump-logs.sh script runs after the build is finished, and
does not affect the status of the build, so it is the perfect
place to include these.