Take a new snapshot for the ABI check.
The google_cloud_cpp_common library ABI has changed, and thus
we need to take a new snapshot of the ABI to receive accurate
reports in the future. I should have taken this snapshot as part
of #703 because I dropped symbols in that PR. In the future, once
the ABI checks start breaking the build this will not happen
anymore.