Use ExternalProject_Add() for gRPC.
This is part of #518. It makes the management of the gRPC
dependency homogenous between Bazel and CMake build, both
will download the same version, as opposed to using a submodule
in CMake and downloads in Bazel. It also should speed up the build
by a few seconds, because the download of a single version is
faster than downloading the full history.