Add (optional) logging for each http request / response (#779)
This is part of the work for #570.

The severity level names were clashing with existing macros in
Win32.  In particular ERROR is a macro defined as '0' in wingdi.h,
which sometimes we indirectly include. I tried using GCP_ as the
prefix.  Hilariously, GCP_ERROR is also defined as 0x8000 is the
same header.