ostream operators for requests and responses. (#782)
This is part of the work for #570. With this change all the
`*Request` and `*Response` classes get a `std::ostream<<`
operator. This will be used to implement logging in a future
version.

Refactor the function to dump binary data for http debug/info
logs to its own file so I can use it for printing request and
response objects.