New LoggingClient to trace requests. (#787)
This fixes #570. With this change the high-level requests
are logged (in a previous PR we added logging of the http
traffic). This will help us troubleshoot problems when
the library may be getting incorrect inputs, as opposed to
the library sending bad requests. This also logs the returned
objects from a http response, to trouble shoot bad parsing in
the library.

Thanks to @mfschwartz for the thorough review.