Use custom initializer for GoogleMock/GoogleTest. (#794)
This fixes #789.  On Windows, with MSVC, when compiling in Debug
mode, the runtime creates a popup dialog if an assertion fails.
That can block a CI build on AppVeyor until it times out, and
because the details of the error are in the dialog, no useful
information is left in the log. With this change the main()
function for our tests sets up the runtime to print assertion
errors to stderr instead.