Fix EnvironmentVariableRestore helper.
We need to make a copy of the char* returned by std::getenv(). On
Windows that pointer is not guaranteed to remain valid after
modifying *other* environment variables. This fixes #796.

Also always use GoogleApplicationDefaultCredentailsHomeVariable to
fetch the HOME (vs APPDATA) variable name.  This was a bit
of cleanup I did while trying to figure out what was broken.