Fix EnvironmentVariableRestore helper. (#799)
Change the `previous_` member variable to `std::string` so we copy
the value returned by `std::getenv()`. On Windows the pointer returned
by `std::getenv()` is not guaranteed to remain valid after modifying
*other* environment variables, thus the need for this copy.

This fixes #796.

Also always use `GoogleApplicationDefaultCredentailsHomeVariable` to
fetch the `HOME` (or `APPDATA`) variable name.  This was a bit
of cleanup I did while trying to figure out what was broken.