Simplify configuration of policies.
This fixes #66. Until now, if you wanted to change one of the
policies in a Table, TableAdmin, or InstanceAdmin object you had
to pass *all* of them in the constructor. With this change you can
just pass the ones that you want to override. That makes life
easier for our users.