Simplify changing retry / backoff policies.
With this change application developers can change just the
RetryPolicy, just the BackoffPolicy, or both. Before it they
needed to provide both policies in the storage::Client()
constructor, even they may want to just just one of them.

Take advantage of this feature in some of the tests, that makes
the tests simpler and demonstrates that it works.