Create scripts for Kokoro Windows build.
Building on Kokoro gives us more bandwidth. The servers have four
cores (or threads), while AppVeyor has only 2, and we get more
parallel builds: Kokoro runs all presubmit builds in parallel.

The disadvantage is that we have to manually do caching, Kokoro does
not make it easy to do that ourselves.

I am getting 35 minute builds with Kokoro, which is worse than the 25
minutes for AppVeyor. I think there is opportunity to improve here, and
the advantage of parallel builds when multiple PRs are running is too
much to ignore.

This commit only introduces the build scripts, after it is committed a
separate CL will enable the build.