Implement storage::ObjectWriteStream (#911)
This is a class derived from std::basic_ostream<char> that uses
libcurl to upload data using chunked enconding. It will be used
in Client::WriteObject(), part of the #557 changes.

Use gunicorn + flask in the integration tests, Werkzeug does not
seem to support chunked transfer encoding, and we need to test
that.

Remove a lot of hard-coded stuff.

Thanks to @houglum for pointing it out.