URGENT: Fix windows builds (#947)
Today was not a good day.  The Windows build broke over the weekend (maybe
late last week, but I did not notice until the weekend).

This morning I could not even troubleshoot the problem because `strawberryperl.com`
was down, and we use that (indirectly) to build openssl on Windows:

https://github.com/Microsoft/vcpkg/issues/3973

I tried to do other stuff, without success. Late in the morning the site was back up.
At first, I thought it was just a matter of adapting to the newer protobuf (3.6.0) that
we got as part of a regular vcpkg update:

https://github.com/Microsoft/vcpkg/commit/c95b6bfdc448295b66da22cfd35d34ed8302aa4d

Well, the changes I made in our `cmake/*` files (included in this PR) did fix the build,
and look like useful changes.  But then the unit tests would not pass because of:

https://github.com/google/protobuf/issues/4773

Which is "Not Good"[tm].  I decided to revert the vcpkg version for now, that got the
tests to compile and pass on my Windows machine.  The programming gods wanted
to toy with this mortal today. Because the `.appveyor.yml` file was not properly clearing
the cache, and the builds on appveyor.com still used the protobuf 3.6 artifacts already
cached there.

BTW, this is probably why the builds did not break immediately: vcpkg reused the cached
versions even other stuff had changed.  Our builds continued to work after that commit
made it through.

I think all the necessary changes are in now.  I will commute home and fix anything else
that is broken when I get there.  May the programming gods be kind to you, oh faithful
reviewer.

* Use targets in add_custom_command.

CMake automatically expands targets to the correct generator
expression in add_custom_command, no need for generator
expressions.

* Set missing paths for PROTOBUF_IMPORT_DIRS.

Depending on your version of CMake you get PROTOBUF_IMPORT_DIRS,
or Protobuf_IMPORT_DIRS, or nothing:

https://cmake.org/cmake/help/v3.10/module/FindProtobuf.html
https://cmake.org/cmake/help/v3.5/module/FindProtobuf.html

Make an educated guess in the file dealing with such nonsense.

* Pin vcpkg to an older version.

The newer version includes protobuf-3.6.0 which is broken on
Windows.