Implement Client::CreateBucket
This is part of the work for #821. It implements the API, the
unit tests, and the code samples. The code samples do *NOT* run
in the CI build yet, because we need to implement DeleteBucket
or the tests will constantly leak buckets. For the same reason,
the integration tests do not call this API either.