Implement Client::PatchBucket. (#1001)
This fixes #823, and computes the Bucket CRUD stuff. It implements the
API, the unit tests, an integration test, code snippets, and runs the
snippets in the CI build.

Thanks to @houglum for suggesting an integration test that changed all
the attributes. In the process I discovered a misnamed field (logPrefix vs.
logObjectPrefix), and that location is not something you can change via a
patch request.