Verify OpenSSL versions match.
This is part of the changes to support multi-threaded applications with
Google Cloud Storage. OpenSSL 1.0.2 (which is still widely deployed)
does not perform any internal locking. Instead, we must provide
callbacks to lock and unlock the OpenSSL data structures. Because we use
OpenSSL via libcurl, we must perform this configuration in the same
version of OpenSSL that libcurl uses. Therefore, linking different
versions of OpenSSL is a bad idea.