Reuse connections in libcurl.
This fixes #1049. We keep a pool of CURL* and CURLM* handles
and reuse them instead of creating a new one for each request.
The handles are separated by endpoint, i.e., there is a
separate pool for /upload/storage/v1 and /storage/v1.

Handles are created on demand, so application with a single
thread will only have one handle of each type.

The pools share a CURLSH* handle, which is an libcurl API to share
connections, SSL state, and DNS resolution results.