Initialize SSL thread callbacks when needed. (#1050)
This fixes #1048.

OpenSSL 1.0.2 (which is still widely deployed) does not perform any
internal locking. Instead, we must provide callbacks to lock and
unlock the OpenSSL data structures. Because we use OpenSSL via
libcurl, we must perform this configuration only if libcurl is
linked against OpenSSL<=1.0.2.

In some environments multiple versions of OpenSSL are available, if we
manage to link the library against a different version than the one used
by libcurl, the callback configurations would not work.  So we
explicitly check the library versions once, and fail to initialize the
library if this is the case.

Use the same SSL library as libcurl on macOS.

On macOS libcurl is linked against the LibreSSL library, we need to
link the same library in our code or we cannot safely use it with
multiple threads.