Improve coverage of all examples. (#1045)
The examples are part of the code we expose to our users, and needs
to be tested and executed.  In this PR:

* We run all examples for the storage client, some were missed in
previous passes.
* We verify that calling the examples without an argument produces a
Usage message.
* We change the some of the Bigtable examples to produce better usage
messages.
* We enable coverage reporting for Cloud Bigtable examples, it was
disabled because we never executed the examples before (see #492).

Using the same table for the full example (aka run-full-example) vs
running each one of the smaller examples does not work in production.
Fixed by using two tables.