Improve errors if the user does not set CMAKE_FLAGS.
Forgetting to set CMAKE_FLAGS is an easy mistake to make. Changed the
README.md file to fail if the user does not set the environment
variable. If the user types `${CMAKE_FLAGS}` without the ending `?`
we also generate an error in the CMake file that explains what is
going on.