Trip log file output in CI builds.
Sometimes we get super large log files, and the CI builds fail
because we printed too much. The common solution is to just print
the top and bottom portions of the log, because the rest is typically
repetitive. This change refactors a few places where we did the trimming
in different ways.