Initial version of CompletionQueue. (#1068)
This is a wrapper around the grpc::CompletionQueue with a nicer
interface. Instead of returning a `void*` when an operation completes
this function invokes a callback provided by the application, passing in
the result of the asynchronous operation.  In this initial version we
support just timers, unary and streaming RPCs will be implemented later.