Link consistent versions of OpenSSL with Bazel (#1088)
This change fixes #1087.  With Bazel builds we were linking against BoringSSL
for the Cloud Bigtable client (via gRPC, which seems to prefer that version with
Bazel), but against the system SSL library for the Cloud Storage client. This
change makes both consistent. With CMake we were already using the system
SSL library for both.

Created an example to first reproduce the problem, and also to show how to
compile and use both.

BoringSSL does not implement BIO_f_base64(), and apparently the EVP_*
functions are "better", so use them when linking against that library.