Implement asynchronous RPCs in completion queue. (#1084)
Add support to asynchronous RPCs in CompletionQueue, and expose it via
the `noex::Table` class, just to demonstrate that it works. This version
does not retry any requests, that will come in a future PR.

Changed the signature of the callback functions to include a
CompletionQueue parameter (the completion queue that completed the
request), that will make it easier to implement retry loops later.