Implement Client::SetBucketIamPolicy().
This fixes #847. I had to fix some of the support classes (IamBindings),
and then implement the API, the usual unit tests, extend the integration
tests to use the new API. This time I added two examples, in line with
the rest of the languages, and the examples are included in the CI
builds.