Implement Client::CopyObject(). (#1101)
This fixes #812. It implements the API, the usual unit tests, adds a
new test case in the integration tests, adds the example, and runs
the example as part of the CI build.

The testbench has a number of changes to validate the encryption keys,
we do not encrypt the data, but verify that reads use the same keys that
were passed on the write operation.

Note that `Objects: copy` only uses the `x-goog-encryption-*` headers,
trying to use different encryption keys for the source and destination
does not work against production, though the documentation is unclear.