Implement Client::GetNotification(). (#1106)
This fixes #1034. It implements the API, the usual unit tests, extends
the integration test to call the API, adds a sample, and runs the sample
as part of the CI build.

Create a new bucket for each run of the integration tests. It is more
predictable to have a clean bucket, and deleting it releases any
notifications.