Implement a latency benchmark.
Measures the latency to create, read, write, and delete objects in the
1MiB range. Part of the work for #572.

Improve the testbench to not keep multiple versions of each object
is versioning is disabled in the Bucket, makes it possible to run
this benchmark against the testbench. In addition, the tesbench should
not perform I/O in the middle of changing its state, as that can create
race conditions with gevent.