Fix leaks with alarms in mock by not creating them.
The semantics of grpc::Alarm are super annoying. Once set, they hold a
reference to the underlying grpc::CompletionQueue, and there does not
seem to be an easy way to release them other than running the event
loop. Unfortunately, we cannot do that in the unit tests, because many
operations are too mocked-out to work with the grpc::CompletionQueue
implementation.