Implement ObjectRewriter. (#1154)
This is part of the work for #816. It introduces a new class to manage
rewrite loops. The `Objects: rewrite` API may require multiple calls
before the (logical) operation completes. The ObjectRewriter offers a
simple member function to run the iteration until it completes. For
applications that want to be notified of progress the class offers a
member function with a callback parameter. Finally the application can
take full control and call `Iterate()` explicitly.

This PR contains integration and unit tests for this class, but we do
not expect application developers to construct it directly. A future PR
will introduce the `Client::RewriteObject` member function to return a
`ObjectRewriter`. In that PR we will introduce the examples and change
both unit and integration tests to use the higher-level API.