Make the install target work and test it.
This fixes #650. To make the install target work we need to install
nlohmann/json.hpp in a directory that does not conflict with other
possible locations for the file. It seems that
google/cloud/storage/internal is a good fit for that.

Had to fix the CMake configuration files to define the CURL::CURL target
if it is not defined. Some older versions of CMake (which we support)
may not set that target and it is easier to write the CMake files
assuming it exists.

Simplified the tests for Bazel as a dependency, they were getting rather
long.

I created a test that verifies the install target by actually installing
the code *and* using the installed code to compile a program. The same
program is compiled with Bazel, but with our code as a dependency.
Finally, the same test program is compiled using make(1), to verify the
generated pkg-config(1) files are useful.