Implement Client::RewriteObject. (#1162)
This fixes #816. The API returns an object that the application can use
to iterate until the rewrite completes. There are helper functions to
iterate until the rewrite completes. That made the unit tests are bit
more complex, because I wanted to test the iteration. The integration
tests also test with CSEK, so they are a bit bigger too.

This PR also introduces an example and runs the example in the CI build.

I expanded the integration tests to ensure we checked iterating over
large objects too.