Fix include rules for headers.
This fixes #1179. Each library should only install its public headers.
Headers used for testing, benchmarks, and examples should not be
installed. The internal headers are more difficult, we should probably
install the minimum set that is used in the implementation of public
types, but that is too fine grained to easily support.