Implements helper classes to retry asynchronous operations (#1141)
This is part of the work for #1002. Implement a retry loop for
asynchronous RPCs. The retry loop uses the retry and backoff policies
that are part of the library, and it introduces a idempotent policy to
deal with partial failures in things like BulkApply.

The semantics of grpc::Alarm are super annoying. Once set, they hold a
reference to the underlying grpc::CompletionQueue, and there does not
seem to be an easy way to release them other than running the event
loop. Unfortunately, we cannot do that in the unit tests, because many
operations are too mocked-out to work with the grpc::CompletionQueue
implementation.