Simplify treatment of protobuf as external project. (#1223)
By default, protobuf decorates its debug libraries with a `d` postfix, I
was compiling both the release and debug versions of protobuf to
guarantee that the version without the `d` was always available. A
simpler solution is to use `protobuf_DEBUG_POSTFIX` to disable the
postfix altogether. As a side-effect we get slightly faster builds.