InsertObject() computes MD5 hashes by default. (#1221)
The application can provide the MD5 hash value if they know it, or
disable the MD5 hash computation explicitly if they want to (but they
are warned against it).

A streaming upload will eventually call the destructor, which was
crashing with exceptions disabled.  That makes it impossible to test it
(and really hard to use it for that matter).