Compute MD5 hashes on downloads. (#1240)
The MD5 hash is computed by default during downloads, and we also
capture the MD5 hash returned in the headers. If exceptions are enabled,
and the two values do not match, an exception is raised when the
download completes.  If exceptions are disabled, the values are captured
as part of the ObjectReadStream, and the application can compare them.

The application can also disable the computation of the hash for a
particular download.

Sometimes the server does not have a MD5, we never report errors in that
case.