Fix function to generate message boundaries.
The function to generate a message boundary was not working if the
initial string was found in the message. I found this because MSVC
issues a warning when the value of string::find() is discarded, which it
was in this case (creating an infinite loop too I think).

This is why we cleanup warnings, because sometimes they show you bugs
that you could not find otherwise.