Do not use std::getenv() on Windows.
On Windows std::getenv() is not thread-safe, the pointer returned might
become invalidated by a concurrent call to _putenv_s() or a related
function:

https://docs.microsoft.com/en-us/cpp/c-runtime-library/reference/getenv-wgetenv?view=vs-2017#remarks

While we only modify the environment in the tests, the applications
using the client libraries might modify the environment for their own
reason.  With this change we use the thread-safe _dupenv_s() function on
Windows.