Verify that mismatched MD5 hashes result in errors.
We need to inject an invalid MD5 hash. When running against the
testbench (and only in that case) we send a magic query parameter that
the testbench parses to inject the fault.

Use a separate file for this integration tests because the
object_integration_test.cc file is too large already. In a future PR we
will refactor more tests to this new file.