Fix warnings about MT vs. MD options with MSVC. (#1285)
MSVC has different runtime support libraries for static vs. dynamically
linked programs. Because gRPC does not support shared libraries on
Windows, we must link statically when we use gRPC. But adding the "use
static runtime" option (/MT or /MTd) creates a warning because the "use
dynamic runtime" option is already in the command-line. The solution is
to *replace* the option instead.