Compute MD5 hash on uploads. (#1284)
With this PR all uploads compute MD5 hashes by default, and compare them
against the hash value reported by the service. If there is a mismatch,
and exceptions are enabled, an exception is reported. Without exceptions
the application needs to check the values themselves.

Note that this validation is on the client side, so the corrupted data
is already stored by GCS by the time the error is detected. This is not
as good as sending the hash value before the data, but we cannot do that
when the data is generated dynamically.