Fix warnings about sscanf(). (#1301)
MSVC warns about using sscanf() because it can be unsafe. We are (I
believe) using it in a safe way, but I do not want to disable the
warning for all files. First because we might use it unsafely in a
future file, and second because the macro used to disable this warning
disables a lot more than just sscanf() warnings. So make a very targeted
removal of the warning only in the files where it is relevant.