Simulate hash failures (#1303)
Implement new option to inject custom headers.

We will use a custom header to request specific faults from the
testbench.  It may also be useful if we ever add a header to the
protocol and the library has not been updated yet.

Verify that mismatched MD5 hashes result in errors.

This PR introduces new integration tests to verify that data corruption
on both uploads and downloads is detected. Or more precisely, that
mismatches in the MD5 hash are detected (or detectable for builds
without exception).

Use a separate file for this integration tests because the
object_integration_test.cc file is too large already. In a future PR we
will refactor more tests to this new file.